refactor: mapInfo를 커스텀 훅으로 관리해 컴포넌트에서 비즈니스 로직 분리 (#108)
